refactor(frontend): rename IDC addresses functions (#6350)
# Motivation

We are going to introduce IDB stores for transactions too. So to keep it
more organized, we rename first the current IDB modules to `*-addresses`
since they refer just to addresses functions.
